+++ src/main/java/com/fpoly/coffeeshop/repository/IUnitRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import com.fpoly.coffeeshop.entity.UnitEntity;
 @Repository
 
-	public interface IUnitRepository extends JpaRepository<UnitEntity, Long> {
+	public interface IUnitRepository extends JpaRepository<UnitEntity, Integer> {
 	
 		public UnitEntity findOneByUnitName(String unitName);
